requirement.txt
Django==5.1.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
PyJWT==2.10.1
djangorestframework-jwt==1.11.0
PyMySQL=1.1.1